Spreadsheet書き込み処理と関数定義を分離
